requirements.txt
streamlit
pandas
numpy
plotly
shapely
fpdf2
kaleido
